Changed tunables from int to double

+++ src/main/java/frc/robot/subsystems/elevator/ElevatorConstants.java
@@ -83,31 +83,31 @@ public class ElevatorConstants {
   public static final int ROTATION_ELEVATOR_MOTOR_CAN_ID = 19;
 
   public static final boolean EXTENSION_INVERTED = false;
-  public static final int EXTENSION_POSITION_PID_F = 0;
+  public static final double EXTENSION_POSITION_PID_F = 0;
   public static final double EXTENSION_POSITION_PID_P = 0.75;
-  public static final int EXTENSION_POSITION_PID_I = 0;
-  public static final int EXTENSION_POSITION_PID_D = 0;
-  public static final int EXTENSION_POSITION_PID_I_ZONE = 0;
+  public static final double EXTENSION_POSITION_PID_I = 0;
+  public static final double EXTENSION_POSITION_PID_D = 0;
+  public static final double EXTENSION_POSITION_PID_I_ZONE = 0;
   public static final double EXTENSION_POSITION_PID_PEAK_OUTPUT = 1.0;
 
   public static final boolean ROTATION_INVERTED = false;
-  public static final int ROTATION_POSITION_PID_F = 0;
+  public static final double ROTATION_POSITION_PID_F = 0;
   public static final double ROTATION_POSITION_PID_P = 1.0;
-  public static final int ROTATION_POSITION_PID_I = 0;
-  public static final int ROTATION_POSITION_PID_D = 0;
-  public static final int ROTATION_POSITION_PID_I_ZONE = 0;
+  public static final double ROTATION_POSITION_PID_I = 0;
+  public static final double ROTATION_POSITION_PID_D = 0;
+  public static final double ROTATION_POSITION_PID_I_ZONE = 0;
   public static final double ROTATION_POSITION_PID_PEAK_OUTPUT = .3;
 
   public static final int PIGEON_ID = 4;
   public static final double EXTENSION_SLOW_PEAK_OUTPUT = 0.15;
   public static final double EXTENSION_MAX_ELEVATOR_VELOCITY = 20000; // theoretical maximum 21305
   public static final double EXTENSION_ELEVATOR_ACCELERATION = EXTENSION_MAX_ELEVATOR_VELOCITY * 10;
-  public static final int EXTENSION_SCURVE_STRENGTH = 0;
+  public static final double EXTENSION_SCURVE_STRENGTH = 0;
 
   public static final double ROTATION_SLOW_PEAK_OUTPUT = 0.15;
   public static final double ROTATION_MAX_ELEVATOR_VELOCITY = 20000; // theoretical maximum 21305
   public static final double ROTATION_ELEVATOR_ACCELERATION = ROTATION_MAX_ELEVATOR_VELOCITY * 10;
-  public static final int ROTATION_SCURVE_STRENGTH = 0;
+  public static final double ROTATION_SCURVE_STRENGTH = 0;
 
   public static final double ELEVATOR_EXTENSION_POSITION_TOLERANCE = .01;
   public static final double DEFAULT_EXTENSION_MOTOR_POWER = 0.5;
